gestion d'erreur read atom case int, chgmt object int -> lint

+++ scheme/include/number.h
@@ -33,7 +33,7 @@ typedef struct num_t {
     union {
 
         double       real;
-        int       integer;
+        long int       integer;
     } this;
 
 } num;
